+++ sources/Connection.hpp
@@ -6,7 +6,7 @@
 /*   By: dsilveri <dsilveri@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/05 11:51:28 by dsilveri          #+#    #+#             */
-/*   Updated: 2023/05/19 12:49:56 by dsilveri         ###   ########.fr       */
+/*   Updated: 2023/05/19 19:39:01 by dsilveri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ class Connection
 		~Connection(void);
 
 		struct pollfd	getPollFd(void);
-		short			getFd(void);
+		int			getFd(void);
 		short			getRevents(void);
 		short			getEvents(void);
 		int				getKeepAliveTimeout(void);
